Moved the date manipulation functions into their own file
