+++ src/lib/components/calendar/CalendarGutter.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { onMount } from "svelte";
-	import { normalizeDateByTimeWithinDay } from "$lib/utils";
+	import { normalizeDateByTimeWithinDay } from "$lib/datemanipulation";
 
 	/**
 	 * The minimum time to display on a given day.
